app/requirements.txt
Django
gunicorn
psycopg2-binary
djangorestframework
markdown
factory_boy
django_factory_boy
django-filter
cancel-token